refactor: 자격증명 관련 테스트
