+++ src/test/java/com/agenticcp/core/domain/cloud/adapter/outbound/aws/account/AwsCredentialManagerTest.java
@@ -1,4 +1,4 @@
-package com.agenticcp.core.domain.cloud.service;
+package com.agenticcp.core.domain.cloud.adapter.outbound.aws.account;
 
 import com.agenticcp.core.common.crypto.EncryptionService;
 import com.agenticcp.core.common.exception.BusinessException;
